requirements1.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.7.0
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
httpie==2.6.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.2
psycopg2-binary==2.9.1
pycodestyle==2.8.0
Pygments==2.10.0
pyparsing==2.4.7
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7